requirements.txt
#aiogram==3.1.1
#python-dotenv==1.0.0

# requirements.txt

# FastAPI and related packages
fastapi
uvicorn
pydantic

# Database
sqlalchemy==1.4.25
psycopg2-binary==2.9.1
alembic==1.7.3

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5

# Kafka
kafka-python==2.0.2

# Redis
redis==3.5.3

# Environment variables
python-dotenv==0.19.0

# Testing
pytest==6.2.5
httpx==0.19.0

# Linting and formatting
flake8==3.9.2
black==21.9b0

# Logging
loguru==0.5.3

# API documentation
swagger-ui-bundle==0.0.9

# Performance monitoring
prometheus-client==0.11.0